+++ api/incubator/src/main/java/io/opentelemetry/api/incubator/logs/ExtendedDefaultLogger.java
@@ -41,6 +41,11 @@ class ExtendedDefaultLogger implements ExtendedLogger {
     }
 
     @Override
+    public ExtendedLogRecordBuilder setException(Throwable throwable) {
+      return this;
+    }
+
+    @Override
     public LogRecordBuilder setTimestamp(long timestamp, TimeUnit unit) {
       return this;
     }