+++ exporters/otlp/testing-internal/src/main/java/io/opentelemetry/exporter/otlp/testing/internal/GrpcSpanExporterBuilderWrapper.java
@@ -12,6 +12,7 @@ import io.opentelemetry.sdk.common.export.RetryPolicy;
 import io.opentelemetry.sdk.trace.data.SpanData;
 import java.time.Duration;
 import java.util.Map;
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
 import javax.annotation.Nullable;
@@ -110,6 +111,18 @@ final class GrpcSpanExporterBuilderWrapper implements TelemetryExporterBuilder<S
   @SuppressWarnings("deprecation") // testing deprecated functionality
   public TelemetryExporterBuilder<SpanData> setChannel(Object channel) {
     builder.setChannel((ManagedChannel) channel);
+    return this;
+  }
+
+  @Override
+  public TelemetryExporterBuilder<SpanData> setServiceClassLoader(ClassLoader serviceClassLoader) {
+    builder.setServiceClassLoader(serviceClassLoader);
+    return this;
+  }
+
+  @Override
+  public TelemetryExporterBuilder<SpanData> setExecutorService(ExecutorService executorService) {
+    builder.setExecutorService(executorService);
     return this;
   }