+++ sdk/trace/src/main/java/io/opentelemetry/sdk/trace/ExtendedSdkTracer.java
@@ -12,16 +12,12 @@ import io.opentelemetry.sdk.trace.internal.TracerConfig;
 
 /** {@link ExtendedSdkTracer} is SDK implementation of {@link ExtendedTracer}. */
 final class ExtendedSdkTracer extends SdkTracer implements ExtendedTracer {
-  // TODO: add dedicated API for updating scope config.
-  @SuppressWarnings("FieldCanBeFinal") // For now, allow updating reflectively.
-  private boolean tracerEnabled;
 
   ExtendedSdkTracer(
       TracerSharedState sharedState,
       InstrumentationScopeInfo instrumentationScopeInfo,
       TracerConfig tracerConfig) {
     super(sharedState, instrumentationScopeInfo, tracerConfig);
-    this.tracerEnabled = tracerConfig.isEnabled();
   }
 
   @Override